redesign database model for stock

+++ controllers/item.controller.ts
@@ -1,9 +1,10 @@
 import { NextFunction, Request, Response } from "express";
-import Item from "../models/item.model";
-import { Item as ItemInterface } from "../interfaces/item.interface";
+import mongoose from "mongoose";
 import { Date } from "../interfaces/date.interface";
-import mongoose, { Types } from "mongoose";
+import { Item as ItemInterface } from "../interfaces/item.interface";
+import Item from "../models/item.model";
 import Stock from "../models/stock.model";
+import { IItem } from "./../models/item.model";
 
 export const addNewItem = async (
   req: Request,
@@ -11,58 +12,45 @@ export const addNewItem = async (
   next: NextFunction
 ) => {
   try {
-    const item: ItemInterface = req.body;
+    const itemData: ItemInterface = req.body;
     // validate item
-    const stock = await Stock.findById(item.stock.toString());
-    if (!stock) throw new Error("No stock found for this id.");
+    const stockId = req.body.stock;
+
+    if (!mongoose.Types.ObjectId.isValid(stockId))
+      throw new Error("Please provide a valid stock id");
+
+    const stock = await Stock.findById(stockId).populate<{ items: IItem[] }>(
+      "items"
+    );
+
+    if (!stock) throw new Error("Stock not found");
 
     // check if item already exists
-    const itemExist = await Item.findOne({
-      name: item.name.toLowerCase().trim(),
-    });
+    const itemExist = stock.items.find(
+      (item) => item.name.toLowerCase() === itemData.name.toLowerCase()
+    );
 
     if (!itemExist) {
-      await Item.create({
-        name: item.name,
-        stock: item.stock,
+      const createdItem = await Item.create({
+        name: itemData.name,
         dates: [
           {
-            date: item.date,
-            quantity: item.quantity,
+            date: itemData.date,
+            quantity: itemData.quantity,
           },
         ],
       });
-      res.status(201).json({ message: "Item added successfully." });
+
+      stock.items.push(createdItem);
+
+      await stock.save();
+
+      res
+        .status(201)
+        .json({ message: "Item added successfully.", data: createdItem });
     } else {
       res.status(404).json({ message: "Item already exists." });
     }
-  } catch (error) {
-    next(error);
-  }
-};
-
-export const getAllItems = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
-    const { stock } = req.query;
-    const queryObject: { stock?: string } = {};
-
-    if (stock && stock !== "undefined") {
-      queryObject.stock = stock.toString();
-    }
-
-    const result = Item.find(queryObject);
-
-    result.sort("name");
-
-    result.populate({ path: "stock" });
-
-    const items = await result;
-
-    res.status(200).json({ data: items });
   } catch (error) {
     next(error);
   }
@@ -75,18 +63,35 @@ export const addNewDateToItem = async (
 ) => {
   try {
     const itemId = req.params.itemId;
+    const stockId = req.params.stockId;
+
     if (!mongoose.Types.ObjectId.isValid(itemId))
-      throw new Error("Please provide a valid id");
+      throw new Error("Please provide a valid item id");
+
+    if (!mongoose.Types.ObjectId.isValid(stockId))
+      throw new Error("Please provide a valid stock id");
 
     const newDate: Date = req.body;
 
-    const item = await Item.findById(itemId);
-    if (!item) throw new Error("Item not found");
+    const stock = await Stock.findById(stockId).populate<{ items: IItem[] }>(
+      "items"
+    );
+    if (!stock) throw new Error("Stock not found");
 
-    item.dates.push(newDate);
-    await item.save();
+    const itemIndex = stock.items.findIndex(
+      (item) => item._id?.toString() === itemId
+    );
 
-    res.status(200).json({ message: "New date added successfully." });
+    if (itemIndex === -1) throw new Error("Item not found");
+
+    stock.items[itemIndex].dates.push(newDate);
+
+    await stock.save();
+
+    res.status(200).json({
+      message: "New date added successfully.",
+      data: stock.items[itemIndex],
+    });
   } catch (error) {
     next(error);
   }
@@ -99,14 +104,30 @@ export const deleteItem = async (
 ) => {
   try {
     const itemId = req.params.itemId;
+    const stockId = req.params.stockId;
+
     if (!mongoose.Types.ObjectId.isValid(itemId))
-      throw new Error("Please provide a valid id");
+      throw new Error("Please provide a valid item id");
+
+    if (!mongoose.Types.ObjectId.isValid(stockId))
+      throw new Error("Please provide a valid stock id");
+
+    const stock = await Stock.findById(stockId);
+    if (!stock) throw new Error("Stock not found");
 
     if (!(await Item.findById(itemId))) throw new Error("Item not found");
 
-    await Item.findByIdAndDelete(itemId);
+    const deletedItem = await Item.findByIdAndDelete(itemId);
 
-    res.status(200).json({ message: "Item deleted successfully." });
+    stock.items.filter(
+      (item) => item._id.toString() !== deletedItem?._id.toString()
+    );
+
+    await stock.save();
+
+    res
+      .status(200)
+      .json({ message: "Item deleted successfully.", data: deletedItem });
   } catch (error) {
     next(error);
   }
@@ -129,7 +150,7 @@ export const updateItemName = async (
     item.name = newName;
     await item.save();
 
-    res.status(200).json({ message: "Item updated successfully." });
+    res.status(200).json({ message: "Item updated successfully.", data: item });
   } catch (error) {
     next(error);
   }
@@ -161,7 +182,7 @@ export const deleteDate = async (
     item.dates = item.dates.filter((_date, index) => index !== dateIndex);
     await item.save();
 
-    res.status(200).json({ message: "Date deleted successfully." });
+    res.status(200).json({ message: "Date deleted successfully.", data: item });
   } catch (error) {
     next(error);
   }
@@ -197,7 +218,9 @@ export const updateDate = async (
     item.dates[dateIndex].quantity = newQuantity;
     await item.save();
 
-    res.status(200).json({ message: "Date quantity updated successfully." });
+    res
+      .status(200)
+      .json({ message: "Date quantity updated successfully.", data: item });
   } catch (error) {
     next(error);
   }
